requirements.txt
pandas>=1.5.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=1.0.0
ax-platform>=1.0.0
tensorboard>=2.8.0 